feat: Implement post test code
