+++ dear_j/blog/tests/schedules.py
@@ -1,0 +1,29 @@
+import pytest
+
+from django import test
+from rest_framework import status
+
+from utils.test import compare as compare_utils
+from utils.test import data as data_utils
+
+
+@pytest.fixture(name="user1")
+def fixture_registered_user1(client: test.Client):
+    user1_data = data_utils.UserData.create_nth_user_data(1)
+    client.post(path="/api/v1/user/registration/", data=user1_data.for_registration, content_type="application/json")
+    return user1_data
+
+
+@pytest.fixture(name="user2")
+def fixture_registered_user2(client: test.Client):
+    user2_data = data_utils.UserData.create_nth_user_data(2)
+    client.post(path="/api/v1/user/registration/", data=user2_data.for_registration, content_type="application/json")
+    return user2_data
+
+@pytest.mark.django_db
+def test_get_post_by_schedule(
+    client: test.Client,
+    user1: data_utils.UserData,
+    user2: data_utils.UserData,
+):
+    pass
