+++ dear_j/pytest.ini
@@ -8,4 +8,4 @@ env =
     GOOGLE_CLIENT_ID="test_client_id"
     GOOGLE_SECRET="test_secret"
     AWS_STORAGE_BUCKET_NAME="dear-j-test-test-bucket"
-python_files = tests.py *_test.py
+python_files = tests.py *_test.py tests/*
